+++ src/physics/vector2D.cpp
@@ -1,22 +1,22 @@
 #include "vector2D.h"
 
-physics::Vector2D::Vector2D(double vX, double vY) {
+physics::Vector2D::Vector2D(float vX, float vY) {
     this->vX = vX;
     this->vY = vY;
 }
 
-double physics::Vector2D::getVX() {
+float physics::Vector2D::getVX() {
     return vX;
 }
 
-void physics::Vector2D::setVX(double vX) {
+void physics::Vector2D::setVX(float vX) {
     this->vX = vX;
 }
 
-double physics::Vector2D::getVY() {
+float physics::Vector2D::getVY() {
     return vY;
 }
 
-void physics::Vector2D::setVY(double vY) {
+void physics::Vector2D::setVY(float vY) {
     this->vY = vY;
 }
